updated network_packet_input.py to use float updated energy_per_packet to value from literature

+++ end-to-end-power-model/energyinput/network_packet_input.py
@@ -1,7 +1,7 @@
 from .generic_input import GenericInput
 
 class NetworkPacketInput(GenericInput):
-    def __init__(self, interface: str, energy_per_packet: int):
+    def __init__(self, interface: str, energy_per_packet: float):
         super().__init__()
         self.interface = interface
         self.energy_per_packet = energy_per_packet
@@ -15,7 +15,7 @@ class NetworkPacketInput(GenericInput):
             packets = int(file.read().strip())
         return packets
 
-    def read_energy(self) -> int:
+    def read_energy(self) -> float:
         current_packets = self.read_packets()
         packets_transmitted = current_packets - self.previous_packets
         self.previous_packets = current_packets
